feat: add use case for get post by id

+++ src/DevSocial.Domain/Repositories/Posts/IPostsReadOnlyRepository.cs
@@ -5,7 +5,7 @@ namespace DevSocial.Domain.Repositories.Posts;
 public interface IPostsReadOnlyRepository
 {
     public Task<List<PostEntitie>> GetAllAsync();
-    public Task<PostEntitie?> GetByIdAsync(int id);
+    public Task<PostEntitie?> GetByIdAsync(long id);
     
     
 }
